requirements.txt
Django==5.0.2
djangorestframework==3.14.0
psycopg2-binary==2.9.9
drf-spectacular==0.27.1
pytest==7.4.4
pytest-django==4.7.0
pytest-cov==4.1.0
black==23.12.1
flake8==7.0.0
isort==5.13.2
psutil==5.9.6 